requirements.txt
streamlit
tensorflow==2.10.1
scikit-learn
pandas
numpy
joblib
requests
python-dateutil
